@@ -16,6 +16,7 @@ public enum ElasticSearchReservedWords
 	SYNONYMS_PATH("synonyms_path"), 
 	ANALYZER("analyzer"), 
 	CUSTOM("custom"), 
+	FIELD_DATA("fielddata"),
 	TOKENIZER("tokenizer"), 
 	WHITESPACE("whitespace"), 
 	LOWERCASE("lowercase"), 
@@ -28,12 +29,17 @@ public enum ElasticSearchReservedWords
 	DATE_FORMATS("date_formats"),
 	FORMAT("format"),
 	STORE("store"), 
-	YES("yes"), 
+	YES("yes"),
+	TRUE("true"),
+	FALSE("false"),
 	INDEX("index"), 
-	NOT_ANALYZED("not_analyzed"), 
+	NOT_ANALYZED("not_analyzed"),
+	ANALYZED("analyzed"),
 	FLOAT("float"), 
 	BOOLEAN("boolean"), 
-	STRING("string"), 
+	STRING("string"),
+	TEXT("text"),
+	COMPLETION("completion"),
 	DOUBLE("double"), 
 	FIELDS("fields"), 
 	MULTI_FIELD("multi_field"), 
@@ -46,10 +52,10 @@ public enum ElasticSearchReservedWords
 	CLUSTER_NAME("cluster.name"), 
 	PATH_DATA("path.data"),
 	PATH_WORK("path.work"),
-	PATH_LOG("path.log"), 
+	PATH_LOG("path.logs"), 
 	PATH_CONF("path.conf"),
-	NUMBER_OF_SHARDS("number_of_shards"), 
-	NUMBER_OF_REPLICAS("number_of_replicas"), 
+	NUMBER_OF_SHARDS("index.number_of_shards"), 
+	NUMBER_OF_REPLICAS("index.number_of_replicas"), 
 	ANALYZER_SIMPLE("simple"), 
 	SYNONYMS_IGNORE_CASE("ignore_case"),
 	SYNONYMS_EXPAND("expand"), 
