@@ -1,53 +1,48 @@
 package org.jai.search.query.impl;
 
-import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
-
-import java.io.IOException;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
 import org.apache.commons.lang.StringUtils;
+import org.apache.lucene.search.join.ScoreMode;
 import org.elasticsearch.action.get.GetResponse;
-import org.elasticsearch.action.mlt.MoreLikeThisRequestBuilder;
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
-import org.elasticsearch.action.suggest.SuggestRequestBuilder;
-import org.elasticsearch.action.suggest.SuggestResponse;
 import org.elasticsearch.index.get.GetField;
-import org.elasticsearch.index.query.AndFilterBuilder;
-import org.elasticsearch.index.query.FilterBuilder;
-import org.elasticsearch.index.query.FilterBuilders;
-import org.elasticsearch.index.query.FilteredQueryBuilder;
-import org.elasticsearch.index.query.NestedFilterBuilder;
-import org.elasticsearch.index.query.OrFilterBuilder;
+import org.elasticsearch.index.query.BoolQueryBuilder;
+import org.elasticsearch.index.query.MoreLikeThisQueryBuilder;
+import org.elasticsearch.index.query.MultiMatchQueryBuilder;
+import org.elasticsearch.index.query.NestedQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.index.query.QueryStringQueryBuilder;
-import org.elasticsearch.index.query.TermFilterBuilder;
+import org.elasticsearch.index.query.SpanTermQueryBuilder;
+import org.elasticsearch.index.query.TermQueryBuilder;
 import org.elasticsearch.index.query.functionscore.FunctionScoreQueryBuilder;
-import org.elasticsearch.index.query.functionscore.ScoreFunctionBuilder;
-import org.elasticsearch.index.query.functionscore.script.ScriptScoreFunctionBuilder;
+import org.elasticsearch.index.query.functionscore.ScoreFunctionBuilders;
+import org.elasticsearch.script.Script;
+import org.elasticsearch.script.ScriptType;
 import org.elasticsearch.search.SearchHit;
 import org.elasticsearch.search.SearchHitField;
-import org.elasticsearch.search.facet.Facet;
-import org.elasticsearch.search.facet.FacetBuilders;
-import org.elasticsearch.search.facet.Facets;
-import org.elasticsearch.search.facet.range.RangeFacet;
-import org.elasticsearch.search.facet.range.RangeFacetBuilder;
-import org.elasticsearch.search.facet.terms.TermsFacet;
-import org.elasticsearch.search.facet.terms.TermsFacet.ComparatorType;
-import org.elasticsearch.search.facet.terms.TermsFacetBuilder;
+import org.elasticsearch.search.aggregations.Aggregation;
+import org.elasticsearch.search.aggregations.AggregationBuilders;
+import org.elasticsearch.search.aggregations.bucket.filters.Filters;
+import org.elasticsearch.search.aggregations.bucket.filters.Filters.Bucket;
+import org.elasticsearch.search.aggregations.bucket.range.RangeAggregationBuilder;
+import org.elasticsearch.search.aggregations.bucket.terms.StringTerms;
+import org.elasticsearch.search.aggregations.bucket.terms.Terms;
+import org.elasticsearch.search.aggregations.bucket.terms.TermsAggregationBuilder;
+import org.elasticsearch.search.rescore.QueryRescorerBuilder;
 import org.elasticsearch.search.rescore.RescoreBuilder;
-import org.elasticsearch.search.rescore.RescoreBuilder.Rescorer;
 import org.elasticsearch.search.sort.SortBuilders;
 import org.elasticsearch.search.suggest.Suggest.Suggestion.Entry.Option;
 import org.elasticsearch.search.suggest.SuggestBuilder;
-import org.elasticsearch.search.suggest.term.TermSuggestion;
-import org.elasticsearch.search.suggest.term.TermSuggestionBuilder;
+import org.elasticsearch.search.suggest.completion.CompletionSuggestionBuilder;
 import org.jai.search.client.SearchClientService;
 import org.jai.search.model.AutoSuggestionEntry;
 import org.jai.search.model.Category;
@@ -85,30 +80,42 @@ public class ProductQueryServiceImpl implements ProductQueryService
                                                                         searchCriteria.getDocumentTypes(), 
                                                                         searchCriteria.getFrom(), 
                                                                         searchCriteria.getSize());
-        requestBuilder.addFields(SearchDocumentFieldName.productQueryFields);
+        for (String field : SearchDocumentFieldName.productQueryFields) {
+        	requestBuilder.addDocValueField(field);	
+		}
+        logger.debug("Search request, with fields: {}" + requestBuilder.toString());
+        
+        //requestBuilder.addFields(SearchDocumentFieldName.productQueryFields);
         
         if(searchCriteria.isRescoreOnSoldOut())
         {
-            Rescorer rescorer = RescoreBuilder.queryRescorer(QueryBuilders.termQuery(SearchDocumentFieldName.SOLD_OUT.getFieldName(), false))
+            QueryRescorerBuilder rescorer = RescoreBuilder.queryRescorer(QueryBuilders.termQuery(SearchDocumentFieldName.SOLD_OUT.getFieldName(), false))
                                                .setQueryWeight(1.0f) //default
                                                .setRescoreQueryWeight(1.5f)
                                                ;
             requestBuilder.setRescorer(rescorer);
         }
         
+        logger.debug("Search request, with score: {}" + requestBuilder.toString());
+        
+        
         if (searchCriteria.hasFilters())
         {
-            AndFilterBuilder andFilterBuilder = getFilterBuilderForSearchCriteria(searchCriteria);
-            requestBuilder.setQuery(QueryBuilders.filteredQuery(queryBuilder, andFilterBuilder));
+        			QueryBuilder andFilterBuilder = getFilterBuilderForSearchCriteria(searchCriteria);
+            requestBuilder.setQuery(QueryBuilders.boolQuery().filter(queryBuilder));
         } else
         {
             requestBuilder.setQuery(queryBuilder);
         }
+        
+        logger.debug("Search request, with filters: {}" + requestBuilder.toString());
 
         if (!searchCriteria.isNoFacets() && searchCriteria.getFacets().size() > 0)
         {
             addFacets(searchCriteria, requestBuilder);
         }
+        
+        logger.debug("Search request, with facets: {}" + requestBuilder.toString());
 
       //Add sorting
         if(searchCriteria.getSortOrder() !=null)
@@ -119,7 +126,11 @@ public class ProductQueryServiceImpl implements ProductQueryService
             requestBuilder.addSort(SortBuilders.scoreSort());
         }
 
-        logger.debug("Executing following search request:" + requestBuilder.internalBuilder().toString());
+//        requestBuilder.setFetchSource(new String[]{"title"}, null);
+        
+        logger.debug("Executing following search request:" + requestBuilder.toString());
+        
+        
         
         SearchResponse searchResponse = requestBuilder.execute().actionGet();
         
@@ -133,7 +144,7 @@ public class ProductQueryServiceImpl implements ProductQueryService
     public Product getProduct(ElasticSearchIndexConfig config, Long productId)
     {
         GetResponse getResponse = searchClientService.getClient().prepareGet(config.getIndexAliasName(), config.getDocumentType(), String.valueOf(productId))
-                                                                 .setFields(SearchDocumentFieldName.productDocumentFields)
+                                                                 .setStoredFields(SearchDocumentFieldName.productDocumentFields)
                                                                  .get();
         if(getResponse.isExists())
         {
@@ -144,6 +155,7 @@ public class ProductQueryServiceImpl implements ProductQueryService
             product.setSoldOut(Boolean.valueOf(getResponse.getField(SearchDocumentFieldName.SOLD_OUT.getFieldName()).getValue().toString()));
             product.setAvailableOn(SearchDateUtils.getFormattedDate(getResponse.getField(SearchDocumentFieldName.AVAILABLE_DATE.getFieldName()).getValue().toString()));
             product.setKeywords(getListFieldValueOrNull(getResponse.getField(SearchDocumentFieldName.KEYWORDS.getFieldName())));
+            if(getResponse.getField(SearchDocumentFieldName.PRICE.getFieldName()) != null)
             product.setPrice(BigDecimal.valueOf(Double.valueOf(getResponse.getField(SearchDocumentFieldName.PRICE.getFieldName()).getValue().toString())));
             product.setBoostFactor(Float.valueOf(getResponse.getField(SearchDocumentFieldName.BOOSTFACTOR.getFieldName()).getValue().toString()));
             GetField catField = getResponse.getField(SearchDocumentFieldName.CATEGORIES_ARRAY.getFieldName());
@@ -170,15 +182,15 @@ public class ProductQueryServiceImpl implements ProductQueryService
     @Override
     public List<AutoSuggestionEntry> getAutoSuggestions(ElasticSearchIndexConfig config, String queryString)
     {
-        TermSuggestionBuilder suggesBuilder = SuggestBuilder.termSuggestion(SearchFacetName.AUTO_SUGGESTION.getCode())
-                                                             .field(SearchDocumentFieldName.KEYWORDS.getFieldName())
+        							SuggestBuilder suggestBuilder = new SuggestBuilder()
+                                                    .addSuggestion(SearchFacetName.AUTO_SUGGESTION.getCode(), 
+                                                     new CompletionSuggestionBuilder(SearchDocumentFieldName.KEYWORDS.getFieldName())
                                                              .analyzer(config.getAutoSuggestionAnalyzerName())
                                                              .size(20)
-                                                             .text(queryString)
+                                                             .text(queryString))
 //                                                             .suggestMode("always")
 //                                                             .stringDistance("ngram")
                                                              ;
-        
 //        CompletionSuggestionBuilder suggesBuilder = new CompletionSuggestionBuilder(SearchFacetName.AUTO_SUGGESTION.getCode())
 //                .field(SearchDocumentFieldName.KEYWORDS.getFieldName())
 //                .analyzer(config.getAutoSuggestionAnalyzerName())
@@ -194,19 +206,20 @@ public class ProductQueryServiceImpl implements ProductQueryService
 //                                                              .text(queryString)
 //                                                              ;
         
-        SuggestRequestBuilder addSuggestion = searchClientService.getClient().prepareSuggest(config.getIndexAliasName())
-                                        .addSuggestion(suggesBuilder);
+        SearchRequestBuilder searchRequestBuilder = searchClientService.getClient().prepareSearch(config.getIndexAliasName())
+                                        .suggest(suggestBuilder);
         
         try
         {
-            logger.debug("Auto Suggestion request is {}", suggesBuilder.toXContent(jsonBuilder().startObject(), null).prettyPrint().string());
-        } catch (IOException e)
+//            logger.debug("Auto Suggestion request is {}", suggestBuilder.toXContent(jsonBuilder().startObject(), null).prettyPrint().string());
+            logger.debug("Auto Suggestion request is {}", suggestBuilder);
+        } catch (Exception e)
         {
             //Do nothing  
             logger.error("Error in to string", e);
         }
 
-        SuggestResponse suggestResponse = addSuggestion.get();
+        SearchResponse suggestResponse = searchRequestBuilder.get();
         
         logger.debug("Auto Suggestion response is {}", suggestResponse);
         
@@ -218,7 +231,8 @@ public class ProductQueryServiceImpl implements ProductQueryService
             {
                 for (Option option : suggestEntry.getOptions())
                 {
-                    int count = ((TermSuggestion.Entry.Option) option).getFreq();
+                	//TODO: Fix this
+                    int count = 0;//((CompletionSuggestion.Entry.Option) option).getHit().getFreq();
                     AutoSuggestionEntry autoSuggestionEntry = new AutoSuggestionEntry(option.getText().string(), count);
                     suggestions.add(autoSuggestionEntry);
                 }
@@ -239,45 +253,55 @@ public class ProductQueryServiceImpl implements ProductQueryService
                                                                                     .setSize(0)
                                                                                     .setQuery(QueryBuilders.matchAllQuery());
         
-        TermsFacetBuilder termsFacetBuilder = FacetBuilders.termsFacet(SearchFacetName.AUTO_SUGGESTION.getCode());
+        
         
         String[] fieldsArray = new String[SearchFacetName.autoSuggestionFields.size() + 1];
         SearchFacetName.autoSuggestionFields.toArray(fieldsArray);
         fieldsArray[fieldsArray.length -1] = SearchDocumentFieldName.KEYWORDS.getFieldName();
 
-        termsFacetBuilder.fields(fieldsArray);
+        MultiMatchQueryBuilder multiMatchQuery = QueryBuilders.multiMatchQuery(queryString);
+        //TODO: fix this, make it multiple field
+//        for (String string : fieldsArray) {
+//        	TermsAggregationBuilder termsFacetBuilder = AggregationBuilders.terms(SearchFacetName.AUTO_SUGGESTION.getCode());
+//        	termsFacetBuilder.field(string);
+//        	searchRequestBuilder.addAggregation(termsFacetBuilder);
+//		}
+        TermsAggregationBuilder termsFacetBuilder = AggregationBuilders.terms(SearchFacetName.AUTO_SUGGESTION.getCode());
+    	termsFacetBuilder.field(SearchFacetName.autoSuggestionFields.get(0));
+    	searchRequestBuilder.addAggregation(termsFacetBuilder);
+        searchRequestBuilder.setQuery(multiMatchQuery);
         
         String lowerCaseQueryString = queryString.toLowerCase();
         String filteredSpecialCharsQueryString = escapeQueryChars(lowerCaseQueryString);
         String matchingRegExString = filteredSpecialCharsQueryString + ".*";
         
-        termsFacetBuilder.regex(matchingRegExString)
-                    .size(20)
-                    .order(ComparatorType.TERM);
+//        termsFacetBuilder.regex(matchingRegExString)
+//                    .size(20)
+//                    .order(ComparatorType.TERM);
 
-        searchRequestBuilder.addFacet(termsFacetBuilder);
+//        searchRequestBuilder.addFacet(termsFacetBuilder);
         
-        logger.debug("Auto Suggestion request is {}", searchRequestBuilder.internalBuilder().toString());
+        logger.debug("Auto Suggestion request is {}", searchRequestBuilder.toString());
         
         SearchResponse searchResponse = searchRequestBuilder.get();
         
         try
         {
-            logger.debug("Auto Suggestion response is {}", searchResponse.toXContent(jsonBuilder().startObject(), null).prettyPrint().string());
-        } catch (IOException e)
+//            logger.debug("Auto Suggestion response is {}", searchResponse.toXContent(jsonBuilder().startObject(), null).prettyPrint().string());
+        	logger.debug("Auto Suggestion response is {}", searchResponse);
+        } catch (Exception e)
         {
             logger.error("Search response tostring error:",e);
         }
         
-        Facets facets = searchResponse.getFacets();
-        for (Facet facet : facets)
+        for (Entry<String, Aggregation> facet : searchResponse.getAggregations().getAsMap().entrySet())
         {
-            if(facet.getName().equals(SearchFacetName.AUTO_SUGGESTION.getCode()))
+            if(facet.getKey().equals(SearchFacetName.AUTO_SUGGESTION.getCode()))
             {
-                TermsFacet termsFacet = (TermsFacet) facet;
-                for (org.elasticsearch.search.facet.terms.TermsFacet.Entry entry : termsFacet.getEntries())
+            	StringTerms aggs = searchResponse.getAggregations().get(facet.getKey());
+                for (org.elasticsearch.search.aggregations.bucket.terms.Terms.Bucket entry : aggs.getBuckets())
                 {
-                    AutoSuggestionEntry term = new AutoSuggestionEntry(entry.getTerm().string(), entry.getCount());
+                    AutoSuggestionEntry term = new AutoSuggestionEntry(entry.getKeyAsString(), Long.valueOf(entry.getDocCount()).intValue());
                     autoSuggestEntries.add(term);
                 }
             }
@@ -288,15 +312,13 @@ public class ProductQueryServiceImpl implements ProductQueryService
     @Override
     public List<Product> findSimilarProducts(ElasticSearchIndexConfig config, String[] fields, Long productId)
     {
-        MoreLikeThisRequestBuilder moreLikeThisRequestBuilder = searchClientService.getClient().prepareMoreLikeThis(config.getIndexAliasName(), config.getDocumentType(), String.valueOf(productId))
-                                                                                                .setField(fields)
-                                                                                                .setMinDocFreq(1)
-                                                                                                .setMinTermFreq(1)
-                                                                                                .setSearchSize(10)
-                                                                                                
-                                                                                                ;
+//    	Morelikethis
+    	QueryBuilder queryBuilder = QueryBuilders.moreLikeThisQuery(fields, null, new MoreLikeThisQueryBuilder.Item[]
+    			{new MoreLikeThisQueryBuilder.Item(config.getIndexAliasName(), config.getDocumentType(), String.valueOf(productId))}).minDocFreq(1).minTermFreq(1);
+        SearchRequestBuilder moreLikeThisRequestBuilder = searchClientService.getClient()
+        		.prepareSearch(config.getIndexAliasName(), config.getIndexAliasName()).setQuery(queryBuilder);
         
-        logger.debug("Executing following search request, fields {}", new Object[]{moreLikeThisRequestBuilder.request().fields()});
+        logger.debug("Executing following search request, fields {}", new Object[]{moreLikeThisRequestBuilder.request()});
         
         SearchResponse searchResponse = moreLikeThisRequestBuilder.get();
 
@@ -317,14 +339,15 @@ public class ProductQueryServiceImpl implements ProductQueryService
         return products;
     }
     
-    private AndFilterBuilder getFilterBuilderForSearchCriteria(SearchCriteria searchCriteria)
+    private BoolQueryBuilder getFilterBuilderForSearchCriteria(SearchCriteria searchCriteria)
     {
-        AndFilterBuilder andFilterBuilder = FilterBuilders.andFilter();
+    	BoolQueryBuilder andFilterBuilder = QueryBuilders.boolQuery();
+//        AndFilterBuilder andFilterBuilder = FilterBuilders.andFilter();
 
         //process single select filters
         for (java.util.Map.Entry<String, String> entry : searchCriteria.getSingleSelectFilters().entrySet())
         {
-            andFilterBuilder.add(getBaseFilterBuilder(entry.getKey(), entry.getValue()));
+            andFilterBuilder.filter(getBaseFilterBuilder(entry.getKey(), entry.getValue()));
         }
 
         //process field value filters
@@ -334,24 +357,24 @@ public class ProductQueryServiceImpl implements ProductQueryService
         if(fieldValueFilters.size() > 0)
         {
             //process multiple entries in or filter
-            final OrFilterBuilder orFilterForFieldValueList = FilterBuilders.orFilter();
+            final BoolQueryBuilder orFilterForFieldValueList = QueryBuilders.boolQuery();
             //processing of fieldValueFilters is following:
             //each list of filter items containing in fieldValueFilters joins with 'OR'
             //each filter item in list joins with 'AND'
             for (Map<String, Object> filterItems : fieldValueFilters)
             {
                 //process all entries in map in and operation.
-                AndFilterBuilder andFilterBuilderForFieldValueMapInList = FilterBuilders.andFilter();
+            	BoolQueryBuilder andFilterBuilderForFieldValueMapInList = QueryBuilders.boolQuery();
                 for (Entry<String, Object> entry : filterItems.entrySet())
                 {
                     //retrieve filter builder
-                    FilterBuilder filterBuilder = getTermFilter(entry.getKey(), String.valueOf(entry.getValue()));
+                	QueryBuilder filterBuilder = getTermFilter(entry.getKey(), String.valueOf(entry.getValue()));
                     //add to and condition
-                    andFilterBuilderForFieldValueMapInList.add(filterBuilder);
+                    andFilterBuilderForFieldValueMapInList.filter(filterBuilder);
                 }
-                orFilterForFieldValueList.add(andFilterBuilderForFieldValueMapInList);
+                orFilterForFieldValueList.filter(andFilterBuilderForFieldValueMapInList);
             }
-            andFilterBuilder.add(orFilterForFieldValueList);
+            andFilterBuilder.filter(orFilterForFieldValueList);
         }
 
         //process multi select filters
@@ -368,78 +391,89 @@ public class ProductQueryServiceImpl implements ProductQueryService
             if(entry.getValue().size() > 1)
             {
                 //process multiple entries in or filter
-                final OrFilterBuilder orFilter = FilterBuilders.orFilter();
+                final BoolQueryBuilder orFilter = QueryBuilders.boolQuery();
                 for (final String filterName : entry.getValue())
                 {
-                    orFilter.add(getBaseFilterBuilder(entry.getKey(), filterName));
+                    orFilter.filter(getBaseFilterBuilder(entry.getKey(), filterName));
                 }
-                andFilterBuilder.add(orFilter);
+                andFilterBuilder.filter(orFilter);
             }
             else
             {
                 //process single entry directly in and filter
-                andFilterBuilder.add(getBaseFilterBuilder(entry.getKey(), entry.getValue().get(0)));
+                andFilterBuilder.filter(getBaseFilterBuilder(entry.getKey(), entry.getValue().get(0)));
             }
         }
         
-        //process resolution/memory facets
-        andFilterBuilder.add(FilterBuilders.nestedFilter(SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), getSpecificationsFacetFilterBuilder(searchCriteria)));
+        ScoreMode mode = ScoreMode.Total;
+		//process resolution/memory facets
+        andFilterBuilder.filter(QueryBuilders.nestedQuery(SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), getSpecificationsFacetFilterBuilder(searchCriteria), mode));
         
         
         //process child product properties
         if(searchCriteria.getProductProperties().size() > 0)
         {
-            FilterBuilder propertyFilterBuilder = null;
+            QueryBuilder propertyFilterBuilder = null;
             if(searchCriteria.getProductProperties().size() > 1)
             {
-                OrFilterBuilder OrPropertyFilterBuilder1 = FilterBuilders.orFilter();
+            	BoolQueryBuilder OrPropertyFilterBuilder1 = QueryBuilders.boolQuery();
                 for (ProductProperty productProperty : searchCriteria.getProductProperties())
                 {
-                    AndFilterBuilder andPropertyTermFilter = FilterBuilders.andFilter(FilterBuilders.termFilter(SearchDocumentFieldName.SIZE.getFieldName(), productProperty.getSize().toLowerCase()),
-                            FilterBuilders.termFilter(SearchDocumentFieldName.COLOR.getFieldName(), productProperty.getColor().toLowerCase()));
-                    OrPropertyFilterBuilder1.add(andPropertyTermFilter);
+                	BoolQueryBuilder andPropertyTermFilter = QueryBuilders.boolQuery()
+                			.filter(
+                			QueryBuilders.termQuery(SearchDocumentFieldName.SIZE.getFieldName(), productProperty.getSize().toLowerCase()))
+                			.filter(
+                            QueryBuilders.termQuery(SearchDocumentFieldName.COLOR.getFieldName(), productProperty.getColor().toLowerCase()))
+                			;
+                    OrPropertyFilterBuilder1.filter(andPropertyTermFilter);
                 }
                 propertyFilterBuilder = OrPropertyFilterBuilder1;
             }
             else
             {
                 ProductProperty productProperty = searchCriteria.getProductProperties().get(0);
-                propertyFilterBuilder = FilterBuilders.andFilter(FilterBuilders.termFilter(SearchDocumentFieldName.SIZE.getFieldName(), productProperty.getSize()),
-                        FilterBuilders.termFilter(SearchDocumentFieldName.COLOR.getFieldName(), productProperty.getColor()));
+                propertyFilterBuilder = QueryBuilders.boolQuery()
+                		.filter(
+                		QueryBuilders.termQuery(SearchDocumentFieldName.SIZE.getFieldName(), productProperty.getSize())
+                		)
+                		.filter(
+                				QueryBuilders.termQuery(SearchDocumentFieldName.COLOR.getFieldName(), productProperty.getColor()))
+                		;
             }
-            FilteredQueryBuilder filteredQueryBuilder = new FilteredQueryBuilder(QueryBuilders.matchAllQuery(), propertyFilterBuilder);
+//            FilteredQueryBuilder filteredQueryBuilder = QueryBuilders.nes(QueryBuilders.matchAllQuery(), propertyFilterBuilder);
+            
             //TODO: config replace
-            andFilterBuilder.add(FilterBuilders.hasChildFilter(ElasticSearchIndexConfig.COM_WEBSITE.getPropertiesDocumentType(), filteredQueryBuilder));
+            andFilterBuilder.filter(QueryBuilders.hasChildQuery(ElasticSearchIndexConfig.COM_WEBSITE.getPropertiesDocumentType(), propertyFilterBuilder, ScoreMode.Total));
         }
         
         //Another approach for specifications, in case faceting not used.
         if(searchCriteria.getSpecifications().size() > 0)
         {
-            FilterBuilder specificationFilterBuilder = null;
+            QueryBuilder specificationFilterBuilder = null;
             specificationFilterBuilder = getSpecificationsFilterBuilder(searchCriteria);
             
-            andFilterBuilder.add(FilterBuilders.nestedFilter(SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), specificationFilterBuilder));
+            andFilterBuilder.filter(QueryBuilders.nestedQuery(
+            		SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), specificationFilterBuilder, ScoreMode.Total));
         }
 
         return andFilterBuilder;
     }
 
-    private FilterBuilder getSpecificationsFilterBuilder(SearchCriteria searchCriteria)
+    private QueryBuilder getSpecificationsFilterBuilder(SearchCriteria searchCriteria)
     {
         //This is used in case you want to search items based on specifications itself.
         
-        FilterBuilder specificationFilterBuilder;
+    	QueryBuilder specificationFilterBuilder;
         List<Specification> specifications = searchCriteria.getSpecifications();
         if(specifications.size() > 1)
         {
-            OrFilterBuilder OrSpecificationFilterBuilder = FilterBuilders.orFilter();
+        	BoolQueryBuilder OrSpecificationFilterBuilder = QueryBuilders.boolQuery();
             for (Specification specification : specifications)
             {
-                FilterBuilder filterBuilder = FilterBuilders.andFilter(FilterBuilders
-                        .termFilter(SearchDocumentFieldName.RESOLUTION.getFieldName(), specification.getResolution()),
-                            FilterBuilders.termFilter(SearchDocumentFieldName.MEMORY.getFieldName(), specification.getMemory()));
-
-                OrSpecificationFilterBuilder.add(filterBuilder);
+            	QueryBuilder filterBuilder = QueryBuilders.spanOrQuery(new SpanTermQueryBuilder(SearchDocumentFieldName.RESOLUTION.getFieldName(), specification.getResolution()))
+            	.addClause(new SpanTermQueryBuilder(SearchDocumentFieldName.MEMORY.getFieldName(), specification.getMemory()));
+            	
+                OrSpecificationFilterBuilder.filter(filterBuilder);
             }
             specificationFilterBuilder = OrSpecificationFilterBuilder;
         }
@@ -447,16 +481,16 @@ public class ProductQueryServiceImpl implements ProductQueryService
         {
             Specification specification = searchCriteria.getSpecifications().get(0);
             
-                    FilterBuilder filterBuilder = FilterBuilders.andFilter(FilterBuilders
-                                                        .termFilter(SearchDocumentFieldName.RESOLUTION.getFieldName(), specification.getResolution()),
-                    FilterBuilders.termFilter(SearchDocumentFieldName.MEMORY.getFieldName(), specification.getMemory()));
+            QueryBuilder filterBuilder = QueryBuilders.boolQuery().must(QueryBuilders
+                                                        .termQuery(SearchDocumentFieldName.RESOLUTION.getFieldName(), specification.getResolution()))
+                    .must(QueryBuilders.termQuery(SearchDocumentFieldName.MEMORY.getFieldName(), specification.getMemory()));
                     
                     specificationFilterBuilder = filterBuilder;
         }
         return specificationFilterBuilder;
     }
     
-    private FilterBuilder getSpecificationsFacetFilterBuilder(SearchCriteria searchCriteria)
+    private QueryBuilder getSpecificationsFacetFilterBuilder(SearchCriteria searchCriteria)
     {
         //This is used in case you want to search based on separate resolution/memory facets
         Map<String, List<String>> multiSelectFilters = searchCriteria.getMultiSelectFilters();
@@ -475,27 +509,27 @@ public class ProductQueryServiceImpl implements ProductQueryService
         }
         if(resolutionFilters.size() == 0 && memoryFilters.size() == 0 )
         {
-            return FilterBuilders.queryFilter(QueryBuilders.matchAllQuery());
+            return QueryBuilders.matchAllQuery();
         }
         
-        AndFilterBuilder andResolutionAndMemoryFilterBuilder = FilterBuilders.andFilter();
+        BoolQueryBuilder andResolutionAndMemoryFilterBuilder = QueryBuilders.boolQuery();
         if(resolutionFilters.size() > 0)
         {
-            OrFilterBuilder OrResolutionFilterBuilder = FilterBuilders.orFilter();
+        	BoolQueryBuilder OrResolutionFilterBuilder = QueryBuilders.boolQuery();
             for (String resolution : resolutionFilters)
             {
-                OrResolutionFilterBuilder.add(FilterBuilders.termFilter(SearchDocumentFieldName.RESOLUTION.getFieldName(), resolution));
+                OrResolutionFilterBuilder.filter(QueryBuilders.termQuery(SearchDocumentFieldName.RESOLUTION.getFieldName(), resolution));
             }
-            andResolutionAndMemoryFilterBuilder.add(OrResolutionFilterBuilder);
+            andResolutionAndMemoryFilterBuilder.filter(OrResolutionFilterBuilder);
         }
         if(memoryFilters.size() > 0)
         {
-            OrFilterBuilder OrMemoryFilterBuilder = FilterBuilders.orFilter();
+        	BoolQueryBuilder OrMemoryFilterBuilder = QueryBuilders.boolQuery();
             for (String memory : memoryFilters)
             {
-                OrMemoryFilterBuilder.add(FilterBuilders.termFilter(SearchDocumentFieldName.MEMORY.getFieldName(), memory));
+                OrMemoryFilterBuilder.filter(QueryBuilders.termQuery(SearchDocumentFieldName.MEMORY.getFieldName(), memory));
             }
-            andResolutionAndMemoryFilterBuilder.add(OrMemoryFilterBuilder);
+            andResolutionAndMemoryFilterBuilder.filter(OrMemoryFilterBuilder);
         }
 //        else if(specifications.size() == 1)
 //        {
@@ -514,26 +548,28 @@ public class ProductQueryServiceImpl implements ProductQueryService
         return andResolutionAndMemoryFilterBuilder;
     }
     
-    private FilterBuilder getBaseFilterBuilder(String facetName, String fieldValue)
+    private QueryBuilder getBaseFilterBuilder(String facetName, String fieldValue)
     {
+    	
         if(facetName.startsWith(SearchFacetName.SEARCH_FACET_TYPE_FACET_PREFIX))
         {
             return getTermFilter(SearchFacetName.CATEGORIES_FIELD_PREFIX + facetName + "." + SearchDocumentFieldName.FACETFILTER.getFieldName(), fieldValue.toLowerCase());
         }
         else if(facetName.startsWith(SearchFacetName.PRODUCT_PRICE_RANGE.getCode()))
         {
-            return FilterBuilders.rangeFilter(SearchDocumentFieldName.PRICE.getFieldName()).includeLower(true).includeUpper(false).from(fieldValue.split("-")[0]).to(fieldValue.split("-")[1]);
+            return QueryBuilders.rangeQuery(SearchDocumentFieldName.PRICE.getFieldName()).includeLower(true).includeUpper(false).from(fieldValue.split("-")[0]).to(fieldValue.split("-")[1]);
         }
         else
         {
-            return FilterBuilders.termFilter(facetName, fieldValue);
+            return QueryBuilders.termQuery(facetName, fieldValue);
         }
 //        return null;
     }
     
-    private TermFilterBuilder getTermFilter(String fieldName, String fieldValue)
+    private TermQueryBuilder getTermFilter(String fieldName, String fieldValue)
     {
-        return FilterBuilders.termFilter(fieldName, fieldValue);
+    	return QueryBuilders.termQuery(fieldName, fieldValue);
+//        return FilterBuilders.termFilter(fieldName, fieldValue);
     }
 
     private void addFacets(SearchCriteria searchCriteria, SearchRequestBuilder requestBuilder)
@@ -542,11 +578,11 @@ public class ProductQueryServiceImpl implements ProductQueryService
         {
             if(SearchFacetName.categoryFacetFields.contains(facetCode))
             {
-                requestBuilder.addFacet(getNewTermsFacet(facetCode, facetCode + "." + SearchDocumentFieldName.FACET.getFieldName()));
+                requestBuilder.addAggregation(getNewTermsFacet(facetCode, facetCode + "." + SearchDocumentFieldName.FACET.getFieldName()));
             }
             else if(SearchFacetName.PRODUCT_PRICE_RANGE.getCode().equals(facetCode))
             {
-                requestBuilder.addFacet(getPriceRangeFacet(facetCode));
+                requestBuilder.addAggregation(getPriceRangeFacet(facetCode));
             }
             else if(SearchFacetName.PRODUCT_PROPERTY_SIZE.getCode().equals(facetCode) || SearchFacetName.PRODUCT_PROPERTY_COLOR.getCode().equals(facetCode))
             {
@@ -561,43 +597,45 @@ public class ProductQueryServiceImpl implements ProductQueryService
                 //TODO, not working
 //                NestedFilterBuilder nestedFilterBuilder = FilterBuilders.nestedFilter(SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), 
 //                        getSpecificationsFacetFilterBuilder(searchCriteria)).join(false);
-                NestedFilterBuilder nestedFilterBuilder = FilterBuilders.nestedFilter(SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), 
-                        FilterBuilders.matchAllFilter()).join(false);
-                TermsFacetBuilder facetFilter = FacetBuilders.termsFacet(facetCode)
+            	NestedQueryBuilder nestedFilterBuilder = QueryBuilders.nestedQuery(SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), 
+                        QueryBuilders.matchAllQuery(), ScoreMode.Total);
+                TermsAggregationBuilder facetFilter = AggregationBuilders.terms(facetCode)
                         .field(SearchDocumentFieldName.SPECIFICATIONS.getFieldName() + "." + SearchDocumentFieldName.RESOLUTION.getFieldName())
-                        .facetFilter(nestedFilterBuilder)
-                        .nested(SearchDocumentFieldName.SPECIFICATIONS.getFieldName());
-                requestBuilder.addFacet(facetFilter);
+//                        .subAggregation(nestedFilterBuilder)
+//                        .nested(SearchDocumentFieldName.SPECIFICATIONS.getFieldName())
+                        ;
+                requestBuilder.addAggregation(facetFilter);
             }
             else if(SearchFacetName.SPECIFICATION_MEMORY.getCode().equals(facetCode))
             {
                 //TODO, not working
-                NestedFilterBuilder nestedFilterBuilder = FilterBuilders.nestedFilter(SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), 
-                        FilterBuilders.matchAllFilter()).join(false);
-                TermsFacetBuilder facetFilter = FacetBuilders.termsFacet(facetCode)
+                NestedQueryBuilder nestedFilterBuilder = QueryBuilders.nestedQuery(SearchDocumentFieldName.SPECIFICATIONS.getFieldName(), 
+                        QueryBuilders.matchAllQuery(), ScoreMode.Total);
+                TermsAggregationBuilder facetFilter = AggregationBuilders.terms(facetCode)
                         .field(SearchDocumentFieldName.SPECIFICATIONS.getFieldName() + "." + SearchDocumentFieldName.MEMORY.getFieldName())
-                        .facetFilter(nestedFilterBuilder)
-                        .nested(SearchDocumentFieldName.SPECIFICATIONS.getFieldName());
-                requestBuilder.addFacet(facetFilter);
+//                        .facetFilter(nestedFilterBuilder)
+//                        .nested(SearchDocumentFieldName.SPECIFICATIONS.getFieldName())
+                        ;
+                requestBuilder.addAggregation(facetFilter);
             }
             
         }
     }
 
-    private RangeFacetBuilder getPriceRangeFacet(String facetCode)
+    private RangeAggregationBuilder getPriceRangeFacet(String facetCode)
     {
-        return FacetBuilders.rangeFacet(facetCode)
+        return AggregationBuilders.range(facetCode)
                                             .field(SearchDocumentFieldName.PRICE.getFieldName())
                                             .addRange(0, 10)
                                             .addRange(10, 20)
                                             .addRange(20, 100);
     }
     
-    protected TermsFacetBuilder getNewTermsFacet(String facetName, String facetField)
+    protected TermsAggregationBuilder getNewTermsFacet(String facetName, String facetField)
     {
-        final TermsFacetBuilder termsFacetBuilder = new TermsFacetBuilder(facetName);
+        final TermsAggregationBuilder termsFacetBuilder = AggregationBuilders.terms(facetName);
         termsFacetBuilder.field(facetField);
-        termsFacetBuilder.order(ComparatorType.TERM);
+        termsFacetBuilder.order(Terms.Order.term(true));
         termsFacetBuilder.size(100);
         return termsFacetBuilder;
     }
@@ -613,52 +651,59 @@ public class ProductQueryServiceImpl implements ProductQueryService
             Product product = new  Product();
             
             product.setId(Long.valueOf(searchHit.getId()));
-            product.setTitle(getFieldValueOrNull(searchHit, SearchDocumentFieldName.TITLE.getFieldName()));
+            product.setTitle(String.valueOf(searchHit.getSource().get(SearchDocumentFieldName.TITLE.getFieldName())));
+            if(getDoubleFieldValueOrNull(searchHit, SearchDocumentFieldName.PRICE.getFieldName()) != null)
             product.setPrice(BigDecimal.valueOf(getDoubleFieldValueOrNull(searchHit, SearchDocumentFieldName.PRICE.getFieldName())));
             product.setSoldOut(Boolean.valueOf(getFieldValueOrNull(searchHit, SearchDocumentFieldName.SOLD_OUT.getFieldName())));
             
             productSearchResult.addProduct(product);
         }
         
-        if(response.getFacets() !=null)
+        if(response.getAggregations() !=null)
         {
-            for (Facet facet : response.getFacets())
+            for (Entry<String, Aggregation> facet : response.getAggregations().asMap().entrySet())
             {
                 FacetResult facetResult = new FacetResult();
-                facetResult.setCode(facet.getName());
+                facetResult.setCode(facet.getKey());
                 
-                if (TermsFacet.TYPE.equals(facet.getType()))
+                
+                
+                if (TermsAggregationBuilder.NAME.equals(facet.getKey()))
                 {
-                    TermsFacet termsFacet = (TermsFacet) facet;
-                    if(termsFacet.getEntries().size() == 0)
+                	Filters aggs = response.getAggregations().get(facet.getKey());
+                	List<? extends Bucket> buckets = aggs.getBuckets();
+//                	TermsAggregator termsFacet = (TermsAggregator) facet;
+                    if(buckets.size() == 0)
                     {
                         continue;
                     }
                     
-                    for (TermsFacet.Entry entry : termsFacet.getEntries())
+                    for (Bucket entry : buckets)
                     {
                         FacetResultEntry facetResultEntry = new FacetResultEntry();
                         
                         //final String term = entry.getTerm().substring(entry.getTerm().indexOf("_") + 1);
-                        facetResultEntry.setTerm(entry.getTerm().string());
-                        facetResultEntry.setCount(entry.getCount());
+                        facetResultEntry.setTerm(entry.getKeyAsString());
+                        facetResultEntry.setCount(entry.getDocCount());
                         
                         facetResult.addFacetResultEntry(facetResultEntry);
                     }
                 }
-                else if(RangeFacet.TYPE.equals(facet.getType()))
+                else if(RangeAggregationBuilder.NAME.equals(facet.getKey()))
                 {
-                    RangeFacet rangeFacet = (RangeFacet) facet;
-                    if(rangeFacet.getEntries().size() == 0)
+                	Filters aggs = response.getAggregations().get(facet.getKey());
+                	List<? extends Bucket> buckets = aggs.getBuckets();
+//                    RangeFacet rangeFacet = (RangeFacet) facet;
+                    if(buckets.size() == 0)
                     {
                         continue;
                     }
                     
-                    for (RangeFacet.Entry entry : rangeFacet.getEntries())
+                    for (Bucket entry : buckets)
                     {
                         FacetResultEntry facetResultEntry = new FacetResultEntry();
-                        facetResultEntry.setTerm(entry.getFromAsString() + " - " + entry.getToAsString());
-                        facetResultEntry.setCount(entry.getCount());
+                        facetResultEntry.setTerm(entry.getKeyAsString());// + " - " + entry.getToAsString());
+                        facetResultEntry.setCount(entry.getDocCount());
                         
                         facetResult.addFacetResultEntry(facetResultEntry);
                     }
@@ -700,9 +745,10 @@ public class ProductQueryServiceImpl implements ProductQueryService
         {
             if(logger.isDebugEnabled())
             {
-                logger.debug("Search response:"+ response.toXContent(jsonBuilder().startObject().prettyPrint(), null).prettyPrint().string());
+                //logger.debug("Search response:"+ response.toXContent(jsonBuilder().startObject().prettyPrint(), null).prettyPrint().string());
+            	logger.debug("Search response: {}", response);
             }
-        } catch (final IOException ex)
+        } catch (final Exception ex)
         {
             //ignore
             logger.error("Error occured while printing search response for debug.", ex);
@@ -722,7 +768,7 @@ public class ProductQueryServiceImpl implements ProductQueryService
         else
         {
             final String filterSpecialCharsQueryString = escapeQueryChars(queryString);
-            final QueryStringQueryBuilder queryStringQueryBuilder = QueryBuilders.queryString(filterSpecialCharsQueryString);
+            final QueryStringQueryBuilder queryStringQueryBuilder = QueryBuilders.queryStringQuery(filterSpecialCharsQueryString);
 
             // Add fields
             queryStringQueryBuilder.field(SearchDocumentFieldName.TITLE.getFieldName(), (float) 0.5)
@@ -740,11 +786,17 @@ public class ProductQueryServiceImpl implements ProductQueryService
         
         if(searchCriteria.isUseBoostingFactor())
         {
-            FunctionScoreQueryBuilder queryBuilder = new FunctionScoreQueryBuilder(matchQueryBuilder);
-            ScoreFunctionBuilder scoreFunctionBuilder = new ScriptScoreFunctionBuilder().script(SearchDocumentFieldName
-                    .getCalculatedScoreScriptForBostFactor());
-            queryBuilder.add(scoreFunctionBuilder);
-            return queryBuilder;
+//            Script script = new Script(SearchDocumentFieldName
+//			        .getCalculatedScoreScriptForBostFactor());
+            Script script = new Script(ScriptType.INLINE, "mustache", SearchDocumentFieldName
+			        .getCalculatedScoreScriptForBostFactor(), new HashMap<>());
+            
+			FunctionScoreQueryBuilder queryBuilder = new FunctionScoreQueryBuilder(matchQueryBuilder,new FunctionScoreQueryBuilder.FilterFunctionBuilder[]{
+                    new FunctionScoreQueryBuilder.FilterFunctionBuilder(
+                    		ScoreFunctionBuilders.scriptFunction(
+                    				script))});
+			//TODO: fix this
+//            return queryBuilder;
         }
 
         return matchQueryBuilder;
